Enhance error handling and refactor configuration management
- Improved error handling in `writeConfig` to differentiate between existing configuration files and I/O errors, enhancing user feedback during initialization.
- Removed unnecessary debug statements and refactored telemetry integration in various components, streamlining the codebase.
- Updated tests to validate changes in error handling and configuration management, ensuring robustness and correctness.
- Enhanced documentation to reflect updates in configuration handling and error management practices.

+++ pkg/store/consume_test.go
@@ -36,8 +36,8 @@ func TestMaterializeTokenRegionWords(t *testing.T) {
 		ids := []uint64{6001, 6002}
 		buf := idx.MaterializeTokenRegionWords(ids)
 		So(len(buf), ShouldEqual, len(ids)*tw)
-		So(buf[0], ShouldEqual, 0xDEADBEEF)
-		So(buf[tw], ShouldEqual, 0xCAFEBABE)
+		So(buf[0], ShouldEqual, uint64(0xDEADBEEF))
+		So(buf[tw], ShouldEqual, uint64(0xCAFEBABE))
 	})
 }
